api de backoffice de cabañas, arreglos en base de datos.

+++ api/db/consultas/imagenes_consultas.py
@@ -1,61 +1,71 @@
+from db.consultas.conexion_base import get_db_connection
 import sqlite3
-from datetime import datetime
+import sys
 import os
 
-RUTA_BD = os.path.abspath("db/hosteria_byteados.db")
+# Añadir el directorio del paquete a sys.path de manera dinámica
+from pathlib import Path
+sys.path.insert(0, str(Path(__file__).parent))
 
-def get_db_connection():
-    '''
-    Devuelve cursor de conexión a la base de datos según la RUTA_BD
-    '''
-    conn = sqlite3.connect(RUTA_BD)
-    conn.row_factory = sqlite3.Row # Para obtener un diccionario en lugar de una tupla
-    return conn
 
-def agregar_imagen(link, descripcion, id_cabania = "NULL"):
+def agregar_imagen(link, descripcion, id_cabania=None):
     '''
     Agrega url de imagen a la base de datos, el parametro id_cabania es opcional. Devuelve True si la operación es exitosa.
     '''
-    conn = get_db_connection() 
-    query = """Insert into Imagenes values
-               (?, ?, ?)"""
-    conn.execute(query,(link, id_cabania, descripcion))
-    cambios = conn.total_changes
-    conn.close()
+    cambios = 0
+    conn = get_db_connection()
+    if conn is not None:
+        query = """Insert into Imagenes values
+                (?, ?, ?)"""
+        conn.execute(query, (link, id_cabania, descripcion))
+        conn.commit()
+        cambios = conn.total_changes
+        conn.close()
     return cambios > 0
+
 
 def eliminar_imagen(link):
     '''
     Elimina la url de imágen ingresada de la base de datos. Devuelve True si la operación es exitosa.
     '''
-    conn = get_db_connection() 
-    query = """Delete from Imagenes
-               Where link = ?"""
-    conn.execute(query, (link,))
-    cambios = conn.total_changes
-    conn.close()
+    cambios = 0
+    conn = get_db_connection()
+    if conn is not None:
+        query = """Delete from Imagenes
+                Where imagen_link = ?"""
+        conn.execute(query, (link,))
+        conn.commit()
+        cambios = conn.total_changes
+        conn.close()
     return cambios > 0
 
-def obtener_imagenes(cabania_id = None):
+
+def obtener_imagenes(cabania_id=None):
     '''
     Obtiene una lista de las imágenes almacenadas en la base de datos -links-,
     si no se especifica la cabania_id se devuelven las fotos sin vinculación a la tabla “cabanias”. 
-    
+
     El formato de devolución es de diccionario:
     {“descripcion”: link, “descripcion2”: link2, …..}
     '''
+    lista_variables = []
+    res = {}
+
     if cabania_id is None:
         condicion_consulta = "is Null"
     else:
         condicion_consulta = "= ?"
+        lista_variables.append(cabania_id)
 
     conn = get_db_connection()
-    query = "SELECT descripcion, imagen_link FROM Imagenes WHERE cabania_id " + condicion_consulta
-    imagenes = conn.execute(query, (cabania_id,)).fetchall()
-    conn.close()
-    
-    res = {}
-    for imagen in imagenes:
-        res[imagen["descripcion"]] = imagen["imagen_link"]
+
+    if conn is not None:
+        query = "SELECT descripcion, imagen_link FROM Imagenes WHERE cabania_id " + condicion_consulta
+        imagenes = conn.execute(query, tuple(lista_variables)).fetchall()
+        conn.close()
+
+        for descripcion, link in imagenes:
+            res[descripcion] = link
 
     return res
+
